First pass of real error messages and one more test
This has changed how parsing is done.
All parsing is done in that "modules" _parse function
and mapping of errors are done in parse.
This allows us to wrap every possible source of failure
and apply a context to it.

Also fixed `RegisterClusterArrayInfo::dim_index` where `and_then_result`
could have been used