Added partial implementation of tests for errors.
There is a problem somewhere, as the PeripheralError in
`register_name_missing` doesn't seem to have a cause.

Also, we have to match differently when we are nested deeper with a
named peripheral. May have to ultimately just have one error for named
and unnamed peripherals/register/etc and implement the display
differently depending on name: Option<String>