Implemented tracking of what field error comes from
Output now looks like this

```text
In field `EN`
Bit range invalid, ParseError
invalid digit found in string
```

Next is to do the same with registers and peripherals, however, one
problem is that if the object doesn't have a name, we have to tell which
"index" it is at. This causes some problems in how to actually send this
info over.

I've added my test for checking the output to make it easier for people
that want to see the progress.