Support `dimableIdentifierType`-style names
As of CMSIS-SVD 1.3.3 the schema allows for `dimableIdentifierType`
names in several places. This change uses the regular expressions from
the CMSIS-SVD 1.3.6 XSD for both `dimableIdentifierType` and
`identifierType`. While there is prose in ARM documentation suggesting
that all names must be valid ANSI C identifiers this is not born out by
the schema constraints.