Use the stable compiler to install `cross`
The `cross` tool is used during testing, but isn't a direct dependency.
Use the stable compiler when installing it instead of the active
toolchain during Travis tests to prevent having to update the MSRV to
support building `cross`.

Also update the MSRV to 1.36.0 in support of crossbeam, a transitive
dependency which requires 1.36.0.