fieldinfo: Properly deal with 32-wide fields with enumerated values
The expression `2_u32.pow(self.bit_range.width)` will overflow when the
bit_range has a width of 32.  As we know that a bit_range width of 32
will mean all enumerated values must be contained in the range anyway
(as long as the EnumeratedValue.value field is u32), we can skip the
checks and avoid dealing with this problem.